ISSUE #3159 - Refactored breadcrumbs; sliptted into two components
